--- faqs/Manas2403.json
@@ -1,4 +1,0 @@
-{
-    "Ques1": "Answer",
-    "Ques2": "Answer"
-}
