--- faqs/RayLikesAnime.json
@@ -1,5 +1,0 @@
-{
-    "what about the security of my data when we submit on your website": "we keep your data private and end to end encrypted with multiple security checks in our specialised directories. Data leaks are impossible.",
-    "what if the file size exceeds 10mb": "you can try compressing the file as much as possible or make multiple submissions from the same form and mentioning it in the description.",
-    "what if we want to submit multiple submissions with the same form": "you can do so. Our software recognizes multiple submissions from the same user as completely distinct with no relation to the other."
-}
