--- faqs/Siddhant-Agarwal4583.json
@@ -1,9 +1,0 @@
-{
-    "what is the max size of storage alloted per account": "max 1 GB of storage is alloted per account",
-    "What type of forms can me stored": "Forms related to any topic can be asked and stored",
-    "Is the data kept private from other users": "Yes the data is kept fully private and secured from other users",
-    "Is there a support for robo check": "Yes there is reCAPTCHA feature to check for robotes",
-    "How do I contact customer service?": "You can go to our official website xyz.com and get contact number of our customer service",
-    "Is there a time limit after which the forms will be deleted": "Yes after 30 days from form submission the forms will be deleted"
-}
-
