--- faqs/aasthaaaa7.json
@@ -1,4 +1,0 @@
-{
-    "Ques1": "How many submissions per person ?",
-    "Ques2": "Where do the forms get saved ?"
-}
