--- faqs/frikinomad.json
@@ -1,4 +1,0 @@
-{
-    "What is data compliance policy?": "we maintian appropriate security of personal data, including protection against unauthorised or unlawful processing",
-    "how can i delete my account and remove data steps?": "Go to the Data & Privacy section -> Scroll to 'Your data & privacy options' -> Select Delete All Data -> Enter Password to confirm"
-}
